+++ .github/ISSUE_TEMPLATE/2-bug-report.yml
@@ -7,19 +7,19 @@ body:
   - type: markdown
     attributes:
       value: |
-        Thank you for submitting a bug report! It helps make Codex better for everyone.
+        Thank you for submitting a bug report! It helps make Conseil better for everyone.
 
-        If you need help or support using Codex, and are not reporting a bug, please post on [codex/discussions](https://github.com/openai/codex/discussions), where you can ask questions or engage with others on ideas for how to improve codex.
+        If you need help or support using Conseil, and are not reporting a bug, please post on [conseil/discussions](https://github.com/openai/conseil/discussions), where you can ask questions or engage with others on ideas for how to improve conseil.
 
-        Make sure you are running the [latest](https://npmjs.com/package/@openai/codex) version of Codex CLI. The bug you are experiencing may already have been fixed.
+        Make sure you are running the [latest](https://npmjs.com/package/@openai/conseil) version of Conseil CLI. The bug you are experiencing may already have been fixed.
 
         Please try to include as much information as possible.
 
   - type: input
     id: version
     attributes:
-      label: What version of Codex is running?
-      description: Copy the output of `codex --version`
+      label: What version of Conseil is running?
+      description: Copy the output of `conseil --version`
   - type: input
     id: model
     attributes:
