+++ .github/ISSUE_TEMPLATE/3-docs-issue.yml
@@ -5,7 +5,7 @@ body:
   - type: markdown
     attributes:
       value: |
-        Thank you for submitting a documentation request. It helps make Codex better.
+        Thank you for submitting a documentation request. It helps make Conseil better.
   - type: dropdown
     attributes:
       label: What is the type of issue?
