+++ .github/workflows/ci.yml
@@ -56,7 +56,7 @@ jobs:
 
       - name: Lint
         run: |
-          pnpm --filter @openai/codex exec -- eslint src tests --ext ts --ext tsx \
+          pnpm --filter @openai/conseil exec -- eslint src tests --ext ts --ext tsx \
             --report-unused-disable-directives \
             --rule "no-console:error" \
             --rule "no-debugger:error" \
