+++ NOTICE
@@ -1,2 +1,2 @@
-OpenAI Codex
+OpenAI Conseil
 Copyright 2025 OpenAI
