+++ package.json
@@ -1,16 +1,16 @@
 {
-  "name": "codex-monorepo",
+  "name": "conseil-monorepo",
   "private": true,
   "description": "Tools for repo-wide maintenance.",
   "scripts": {
-    "release": "pnpm --filter @openai/codex run release",
+    "release": "pnpm --filter @openai/conseil run release",
     "format": "prettier --check *.json *.md .github/workflows/*.yml",
     "format:fix": "prettier --write *.json *.md .github/workflows/*.yml",
-    "build": "pnpm --filter @openai/codex run build",
-    "test": "pnpm --filter @openai/codex run test",
-    "lint": "pnpm --filter @openai/codex run lint",
-    "lint:fix": "pnpm --filter @openai/codex run lint:fix",
-    "typecheck": "pnpm --filter @openai/codex run typecheck",
+    "build": "pnpm --filter @openai/conseil run build",
+    "test": "pnpm --filter @openai/conseil run test",
+    "lint": "pnpm --filter @openai/conseil run lint",
+    "lint:fix": "pnpm --filter @openai/conseil run lint:fix",
+    "typecheck": "pnpm --filter @openai/conseil run typecheck",
     "changelog": "git-cliff --config cliff.toml --output CHANGELOG.ignore.md $LAST_RELEASE_TAG..HEAD",
     "prepare": "husky",
     "husky:add": "husky add"
@@ -44,7 +44,7 @@
     ".github/workflows/*.yml": "prettier --write",
     "**/*.{js,ts,tsx}": [
       "prettier --write",
-      "pnpm --filter @openai/codex run lint",
+      "pnpm --filter @openai/conseil run lint",
       "cd codex-cli && pnpm run typecheck"
     ]
   },
